# Experiment: avoid copying data back to write buffer from pending data