… src/Storages/ExternalStream/ExternalStreamCounter.h