PR checklist:
- Did you run ClangFormat ?
- Did you separate headers to a different section in existing community code base ?
- Did you surround `proton: starts/ends` for new code in existing community code base ?

Please write user-readable short description of the changes:

This PR implements a solution that avoids copying data from `pending_data` back to `wb`. The idea is use `swap` instead. When it detects there are data in `pending_data`, after `wb` reset (calling `next()`), it swaps `pending_data` and `wb`, so there is no extra `memcpy`.