@zliang-min 2024-06-08
I did a simple performance test:

1. Created a stream called `users`, and inserted 5 million records into it. Each row, after encoded into JSON, its is about 80 bytes, give or take.
2. Created a external stream called `foo_json`.
3. Insert data into `foo_json` by selecting all data from `users` by
    ```sql
    INSERT INTO foo_json (name, favorite_number, favorite_color)
    SETTINGS kafka_max_message_size = 100
    SELECT
      name, favorite_number, favorite_color
    FROM
      table(users)
    ```
    This SQL will make it hit the pending_data copy logic triggered for every row. So that we can tell the differences between before and after the change. The SQL was run 5 rounds for each.

### Result before the change
#### Round 1
0 rows in set. Elapsed: 34.039 sec. Processed 5.00 million rows, 228.45 MB (146.89 thousand rows/s., 6.71 MB/s.)

#### Round 2
0 rows in set. Elapsed: 34.026 sec. Processed 5.00 million rows, 228.45 MB (146.95 thousand rows/s., 6.71 MB/s.)

#### Round 3
0 rows in set. Elapsed: 34.538 sec. Processed 5.00 million rows, 228.45 MB (144.77 thousand rows/s., 6.61 MB/s.)

#### Round 4
0 rows in set. Elapsed: 34.026 sec. Processed 5.00 million rows, 228.45 MB (146.95 thousand rows/s., 6.71 MB/s.)

#### Round 5
0 rows in set. Elapsed: 34.523 sec. Processed 5.00 million rows, 228.45 MB (144.83 thousand rows/s., 6.62 MB/s.)

### Result after the change
#### Round 1
0 rows in set. Elapsed: 35.021 sec. Processed 5.00 million rows, 228.45 MB (142.77 thousand rows/s., 6.52 MB/s.)

#### Round 2
0 rows in set. Elapsed: 35.028 sec. Processed 5.00 million rows, 228.45 MB (142.74 thousand rows/s., 6.52 MB/s.)

#### Round 3
0 rows in set. Elapsed: 35.524 sec. Processed 5.00 million rows, 228.45 MB (140.75 thousand rows/s., 6.43 MB/s.)

#### Round 4
0 rows in set. Elapsed: 35.023 sec. Processed 5.00 million rows, 228.45 MB (142.76 thousand rows/s., 6.52 MB/s.)

#### Round 5
0 rows in set. Elapsed: 35.526 sec. Processed 5.00 million rows, 228.45 MB (140.74 thousand rows/s., 6.43 MB/s.)

### Summary
From the above results, the performance was even better with the original implementation (with `memcpy`).