@zliang-min 2024-06-10
I have done another round a test with 1,000,000 records, and the records' sizes are around 1KiB ( something between 900 bytes to 1500 bytes ).

The results are still look similar. The performance is better without the changes in this PR
```sql
INSERT INTO users_big_json SELECT
  * EXCEPT _tp_time
FROM
  table(users_big)
LIMIT 1000000
SETTINGS kafka_max_message_size = 1800
```

### Results without the changes ( develop branch )

#### Round 1
0 rows in set. Elapsed: 60.544 sec. Processed 1.01 million rows, 1.25 GB (16.71 thousand rows/s., 20.71 MB/s.)

#### Round 2
0 rows in set. Elapsed: 58.045 sec. Processed 1.01 million rows, 1.25 GB (17.43 thousand rows/s., 21.60 MB/s.)

#### Round 3
0 rows in set. Elapsed: 58.039 sec. Processed 1.01 million rows, 1.25 GB (17.44 thousand rows/s., 21.60 MB/s.)

### Results with the changes ( this PR's branch )

#### Round 1
0 rows in set. Elapsed: 63.542 sec. Processed 1.01 million rows, 1.25 GB (15.93 thousand rows/s., 19.73 MB/s.)

#### Round 2
0 rows in set. Elapsed: 64.046 sec. Processed 1.01 million rows, 1.25 GB (15.80 thousand rows/s., 19.58 MB/s.)

#### Round 3
0 rows in set. Elapsed: 62.545 sec. Processed 1.01 million rows, 1.25 GB (16.18 thousand rows/s., 20.05 MB/s.)